chore: move feedback types to own file
