+++ packages/playground/src/api/documents/shared.ts
@@ -2,7 +2,6 @@ import {
   CogniteExternalId,
   CogniteInternalId,
   ExternalId,
-  FilterQuery,
   ItemsWrapper,
 } from '@cognite/sdk-core';
 
@@ -330,61 +329,6 @@ export type StringToStringMap = {
 export type StringToAnyMap = {
   [key: string]: string;
 };
-
-export interface DocumentFeedbackAggregateRequest {
-  field: AggregateField;
-}
-
-export type AggregateField = 'action' | 'status';
-
-export interface AggregateGroup {
-  value: string;
-  count: number;
-}
-
-export interface DocumentFeedbackAggregateResponse {
-  field: AggregateField;
-  groups: AggregateGroup[];
-  total: number;
-}
-
-export type DocumentId = CogniteInternalId;
-
-export interface DocumentFeedbackCreateItem {
-  documentId: DocumentId;
-  label: FeedbackLabel;
-  action: FeedbackAction;
-  reporterInfo?: ReporterInfo;
-}
-
-export interface FeedbackLabel {
-  externalId: CogniteExternalId;
-}
-
-export type FeedbackAction = 'ATTACH' | 'DETACH';
-
-export type ReporterInfo = string;
-
-export type FeedbackStatus = 'CREATED' | 'ACCEPTED' | 'REJECTED' | 'STALE';
-
-export interface FeedbackQueryParameters extends FilterQuery {
-  status?: FeedbackStatus;
-}
-
-export interface DocumentFeedback {
-  documentId: DocumentId;
-  label: FeedbackLabel;
-  action: FeedbackAction;
-  feedbackId: number;
-  reporterInfo?: ReporterInfo;
-  createdAt: string;
-  reviewedAt?: string | null;
-  status: FeedbackStatus;
-}
-
-export interface FeedbackId {
-  id: number;
-}
 
 export interface DocumentPreviewTemporaryLink {
   temporaryLink: string;
