+++ packages/playground/src/types.ts
@@ -8,6 +8,7 @@ import {
 
 export * from './api/documents/shared';
 export * from './api/documents/classifierTypes';
+export * from './api/documents/feedbackTypes';
 
 // This file is here mostly to allow apis to import { ... } from '../../types';
 // Overriding types should probably be done in their respective API endpoint files, where possible
